Create psuedo-random-snippet.js
This JavaScript code can be used at the beginning of an online experiement using jsPsych to pseudorandomize a stimuli list (fillers plus control/experimental) according to the desired parameters. As is, the script has a burn-in of 7 filler items at the beginning, and requires at least 1 filler between control/experimental items.